requirements.txt
numpy==1.16.1
Pillow==3.1.2
matplotlib==1.5.1
opencv-python==3.4.1.15
scipy==1.2.3
psutil==3.4.2
tensorflow==1.8.0
